Refactor Payment links API
No actual changes here.
