+++ mollie/api/client.py
@@ -10,7 +10,7 @@ from requests_oauthlib import OAuth2Session
 from urllib3.util import Retry
 
 from .error import RequestError, RequestSetupError
-from .resources import Chargebacks, Methods, Payments, Refunds
+from .resources import Chargebacks, Methods, PaymentLinks, Payments, Refunds
 from .resources.clients import Clients
 from .resources.customer_mandates import CustomerMandates
 from .resources.customer_payments import CustomerPayments
@@ -20,7 +20,6 @@ from .resources.invoices import Invoices
 from .resources.onboarding import Onboarding
 from .resources.orders import Orders
 from .resources.organizations import Organizations
-from .resources.payment_links import PaymentLinks
 from .resources.permissions import Permissions
 from .resources.profile_chargebacks import ProfileChargebacks
 from .resources.profile_methods import ProfileMethods
