+++ mollie/api/resources/__init__.py
@@ -1,6 +1,7 @@
 from .captures import PaymentCaptures, SettlementCaptures
 from .chargebacks import Chargebacks, PaymentChargebacks
 from .methods import Methods
+from .payment_links import PaymentLinks
 from .payments import Payments
 from .refunds import OrderRefunds, PaymentRefunds, Refunds, SettlementRefunds
 
@@ -11,6 +12,7 @@ __all__ = [
     Payments,
     PaymentCaptures,
     PaymentChargebacks,
+    PaymentLinks,
     PaymentRefunds,
     Refunds,
     SettlementCaptures,
