+++ mollie/api/resources/payment_links.py
@@ -1,4 +1,3 @@
-from ..error import IdentifierError
 from ..objects.payment_link import PaymentLink
 from .base import ResourceBase, ResourceCreateMixin, ResourceGetMixin, ResourceListMixin
 
@@ -12,9 +11,6 @@ class PaymentLinks(ResourceBase, ResourceCreateMixin, ResourceGetMixin, Resource
     def get_resource_object(self, result):
         return PaymentLink(result, self.client)
 
-    def get(self, payment_link_id, **params):
-        if not payment_link_id or not payment_link_id.startswith(self.RESOURCE_ID_PREFIX):
-            raise IdentifierError(
-                f"Invalid payment ID: '{payment_link_id}'. A payment ID should start with '{self.RESOURCE_ID_PREFIX}'."
-            )
+    def get(self, payment_link_id: str, **params):
+        self.validate_resource_id(payment_link_id, "payment link ID")
         return super().get(payment_link_id, **params)
